update entity package: add PlayerBot entity, move entities to appropriate subpackages
