+++ src/main/java/self/izouir/bitkionline/entity/egg/EggType.java
@@ -1,4 +1,4 @@
-package self.izouir.bitkionline.entity;
+package self.izouir.bitkionline.entity.egg;
 
 public enum EggType {
     WEAK, STRONG, HOLY
