+++ src/main/java/self/izouir/bitkionline/entity/player/PlayerBot.java
@@ -1,0 +1,27 @@
+package self.izouir.bitkionline.entity.player;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Table(name = "player_bots")
+@Entity
+public class PlayerBot {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "player_id")
+    private Long playerId;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "last_bot_state")
+    private String lastBotState;
+}
